+++ packages/veritone-react-common/src/components/SourceManagement/story.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import NullState from './Nullstate';
+import NullState from './NullState';
 import SourceTileView from './SourceTileView';
 
 // a mock return result on a source from graphql
@@ -9,6 +9,7 @@ const sourceResult = {
     source: {
       id: '666',
       name: 'KWOL--FM',
+      isLivestream: true,
       createdDateTime: '2014-12-01T18:17:20.675Z',
       modifiedDateTime: '2015-12-01T18:17:20.675Z',
       thumbnailUrl: 'https://image.flaticon.com/icons/svg/25/25305.svg',
@@ -20,6 +21,7 @@ const sourceResult = {
       sourceType: {
         id: '1',
         name: 'Audio',
+        isLive: true,
         sourceSchema: {
           id: 'schemaId1',
           definition: {