+++ packages/veritone-react-common/src/components/formComponents/DateTimePicker.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import Today from '@material-ui/icons/Today';
 import dateFns from 'date-fns';
+import { isString } from 'lodash';
 import TextField from '@material-ui/core/TextField';
 import { instanceOf, func, shape, string, bool } from 'prop-types';
 
@@ -19,8 +20,17 @@ export default class DateTimePicker extends React.Component {
   };
 
   handleDateChange = ({ target }) => {
+    const newDate = target.value;
+
+    if (
+      !dateFns.isValid(new Date(newDate)) ||
+      dateFns.getYear(newDate) > 9999
+    ) {
+      return;
+    }
+
     this.props.input.onChange(
-      consolidate(target.value, getTimeString(this.props.input.value))
+      consolidate(newDate, getTimeString(this.props.input.value))
     );
   };
 
@@ -96,16 +106,18 @@ TimeSelector.propTypes = {
 };
 
 const TimeZoneField = ({ value }) => {
-  return value ? (
-    <TextField
-      className={styles.dateTimeTZ}
-      value={value}
-      InputProps={{
-        disableUnderline: true
-      }}
-      disabled
-    />
-  ) : null;
+  return (
+    value && (
+      <TextField
+        className={styles.dateTimeTZ}
+        value={value}
+        InputProps={{
+          disableUnderline: true
+        }}
+        disabled
+      />
+    )
+  );
 };
 
 TimeZoneField.propTypes = {
@@ -126,15 +138,12 @@ function getTimeString(date) {
 
 function getTimeZone(date) {
   let tzDate = date;
+  if (isString(tzDate)) {
+    tzDate = new Date();
+  }
   if (dateFns.isDate(tzDate)) {
     const tzMatch = tzDate.toTimeString().match(/\(([^)]+)\)$/);
-    if (tzMatch && tzMatch.length > 1) {
-      const tzParts = tzMatch[1].split(' ');
-      if (tzParts.length > 1) {
-        return tzParts.map(part => part[0]).join('');
-      }
-      return tzMatch[1];
-    }
+
+    return tzMatch ? tzMatch[1] : '';
   }
-  return '';
 }